bot/requirements.txt
discord.py
python-dotenv
psutil
requests
docker